Blood Pressure Threshold Method for the app

+++ Hart_Check_Official/Models/BloodPressureThreshold.cs
@@ -8,8 +8,8 @@ namespace Hart_Check_Official.Models
         public int thresholdID { get; set; }
         public int patientID { get; set; }
         public int doctorID { get; set; }
-        public int systolicLevel { get; set; }
-        public int diastolicLevel { get; set; }
+        public double systolicLevel { get; set; }
+        public double diastolicLevel { get; set; }
 
         public Patients patients { get; set; }
         public HealthCareProfessional doctor { get; set; }
